New folder/requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.8.0
python-dotenv==1.1.1
# Optional for Postgres deployments
psycopg2-binary>=2.9
